STATISTIK-607 Total- och summakolumn borttagen för resten av rapporterna.

+++ statistics-web/src/main/java/se/inera/statistics/web/service/DiagnosisSubGroupsConverter.java
@@ -102,7 +102,7 @@ public class DiagnosisSubGroupsConverter {
     private int sum(List<Integer> numbers) {
         int sum = 0;
         for (Integer number : numbers) {
-            sum += number.intValue();
+            sum += number;
         }
         return sum;
     }
